Add new levels and visible_name
New leves added: Service, Endpoint, Software, Server, Datacenter

Add the option to show a short name in the UI. This should be declared
as Metadata.VisibleName.
Graffiti use Name and Type as the primary key, therefore, we must use
long names to differenciate between nodes of the same Type.
With this option we can show a more useful name while keeping the
difference for the primary key