Mark nodes with alarms
If some node has an alarm (Metadata.Alarms[].{Name: String, Level: Int}),
mark its circle in yellow (warning) or red (critical).
Level is a value expected between 0 and 10.
alarmCriticalThreshold defines the threshold between warning and
critical.

Also, all nodes with an alarm or nodes having children with an alarm,
get a badge.
If the higher level of the node/children is "warning", the badge is
an exclamation mark.
If the higher level is "critical", get a broken hearth.

The idea behind this implementation is to push alarms from the
monitoring system to Graffite and the able to track the impact (from
children to parent).