Create M-N realtionships using link type ownership_X
Nodes could only have one parent, which leads us to some difficulties to
represent some relationships.
For example, endpoints should have Software, but VMs also have Software.
This double ownership is not allowed by Graffiti.

This workaround is based on using a different type of link type, those
starting with "ownership_".
Another assumption is also added, only nodes coming from a Service node
should be drawn. This is an importante change, we are basing the
representation on the Service nodes.

With this implementation, parent-child realtionship works normal.
Other nodes, connected directly to root, are hidden by default.
When a child of a "Service" node is drawn, the app checks if it has some
"pseudo children" (children with link type "ownership_X") and those are
displayed too (this pseudo children scan is recursive, displaying all
the nodes connected by this "ownership_X" link type).

Groups are not working correctly.
If we have several nodes gruped, and one of them is a pseudo children,
the group icon is shown, instead of just the node.
If we collapse the view after the group has been shown, the icon remains
drawn in the screen.

Another problem is how these pseudo children are drawn.
We have to think the app have one root node and it draws following the
parent-child relationships.
Those pseudo children nodes are hanging directly from the root node,
therefore they are going to be displayed at the borders of our main
"Service" tree.
Maybe this could be fixed, centering node types without parents (except
root).