Center pseudochildren nodes. Always tag. Big group size
Pesudo children nodes are displayed at the edges of the Services tree,
because they are hanging directly from the root node.
Introduce a new function that analyze each level and if it finds only
root-children nodes, center them in the screen.
It does not try to sort the nodes in any way, therefore could be
a better placement of nodes to reduce links crosses.

Groups are not correctly displayed currently, so increase the value to
25 to avoid showing them, at least for firt demo.

Introduce a special tag "_always".
Nodes with this tag will be always shown, no matter which tag is
selected.