Added timeline to go back in time
Add a time picker to set the G.At() Gremlin filter to be able to see the
graph in a particular point in time.

To handle date the <MuiPickersUtilsProvider> provider has been
configured on top of the App, this has tabulated all the code.
The only change is to add the <DateTimePicker> has been added
at the right side of filters, and the default filter has changed
to "Now".

The "Namespace" filter has been removed but leaved as reference.